+++ changelog.md
@@ -1,3 +1,11 @@
 # Changelog
 
-Rewrote './templates/sequence-data/context.txt' as 'readme.md'
+**./src/common/sub-traverse/errors/**
+* Renamed 'tr-error-txt.js' to 'traverse-error-text.js
+
+---
+
+**Requirement Paths**
+* Updated following:
+	* ./src/common/sub-traverse/traverse-help-tasks.js
+	* ./src/common/sub-traverse/traverse-error-display.js